# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brasil a Vera API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Parlamentar {
  id: string;
  nome: string;
  casa: string;
  partidoSigla: string;
  partidoNome?: string;
  uf: string;
  /** @nullable */
  urlFoto?: string | null;
  legislatura: number;
  trustLevel: string;
  sourceUrl: string;
}

export interface ParlamentarDetalhe {
  id: string;
  sourceId?: string;
  nome: string;
  /** @nullable */
  nomeCivil?: string | null;
  casa: string;
  partidoSigla: string;
  partidoNome?: string;
  uf: string;
  /** @nullable */
  urlFoto?: string | null;
  situacaoMandato?: string;
  legislatura: number;
  trustLevel: string;
  sourceUrl: string;
}

export interface ParlamentaresFiltros {
  partidos: string[];
  ufs: string[];
}

export interface VotoRecente {
  voto: string;
  votacaoId: string;
  /** @nullable */
  votacaoSourceId?: string | null;
  dataHora: string;
  descricao: string;
  orgao: string;
  casa: string;
  aprovada: boolean;
}

export interface ProposicaoResumo {
  id: string;
  tipo: string;
  numero: number;
  ano: number;
  ementa: string;
  situacao: string;
}

export interface CategoriaGasto {
  categoriaDescricao: string;
  total: string;
  n: number;
}

export interface GastosResumo {
  totalGeral: string;
  totalRegistros: number;
  topCategorias: CategoriaGasto[];
  ano: number;
}

export interface VotacaoAlinhamento {
  votacaoId: string;
  dataHora: string;
  descricao: string;
  voto: string;
  orientacao: string;
}

export interface AlinhamentoResult {
  /** @nullable */
  partidoSigla?: string | null;
  /** @nullable */
  percentual?: number | null;
  total: number;
  alinhados: number;
  divergentes: number;
  amostraInsuficiente: boolean;
  topDivergencias: VotacaoAlinhamento[];
  topConvergencias: VotacaoAlinhamento[];
}

export interface AfinidadeVoto {
  parlamentarId: string;
  nome: string;
  partidoSigla: string;
  uf: string;
  casa: string;
  /** @nullable */
  urlFoto?: string | null;
  votosCoincidentes: number;
  totalVotosEmComum: number;
  percentualAfinidade: number;
}

export interface Votacao {
  id: string;
  sourceId: string;
  casa: string;
  /** @nullable */
  proposicaoId?: string | null;
  dataHora: string;
  descricao: string;
  orgao: string;
  aprovada: boolean;
  votosSim: number;
  votosNao: number;
  abstencoes: number;
  trustLevel: string;
  sourceUrl: string;
}

export interface VotoNominal {
  parlamentarId: string;
  nome: string;
  partidoSigla: string;
  uf: string;
  casa: string;
  /** @nullable */
  urlFoto?: string | null;
  voto: string;
}

export interface OrientacaoBancada {
  partidoSigla: string;
  orientacao: string;
}

export interface VotacaoDetalhe {
  id: string;
  sourceId: string;
  casa: string;
  /** @nullable */
  proposicaoId?: string | null;
  dataHora: string;
  descricao: string;
  orgao: string;
  aprovada: boolean;
  votosSim: number;
  votosNao: number;
  abstencoes: number;
  /** @nullable */
  ausentes?: number | null;
  trustLevel: string;
  sourceUrl: string;
  votos: VotoNominal[];
  orientacoes?: OrientacaoBancada[];
}

export interface VotacoesFiltros {
  anos: number[];
}

export interface Proposicao {
  id: string;
  tipo: string;
  numero: number;
  ano: number;
  ementa: string;
  /** @nullable */
  ementaDetalhada?: string | null;
  situacao: string;
  /** @nullable */
  regime?: string | null;
  trustLevel: string;
  sourceUrl: string;
}

export interface Tema {
  codigoTema: number;
  nomeTema: string;
}

export interface Autor {
  /** @nullable */
  parlamentarId?: string | null;
  nome: string;
  tipoAutoria: string;
}

export interface Tramitacao {
  id: string;
  data: string;
  orgao: string;
  descricaoResumida: string;
  /** @nullable */
  descricaoCompleta?: string | null;
  /** @nullable */
  situacaoResultante?: string | null;
}

export interface ProposicaoDetalhe {
  id: string;
  tipo: string;
  numero: number;
  ano: number;
  ementa: string;
  /** @nullable */
  ementaDetalhada?: string | null;
  situacao: string;
  /** @nullable */
  regime?: string | null;
  trustLevel: string;
  sourceUrl: string;
  /** @nullable */
  sourceUrlCamara?: string | null;
  /** @nullable */
  sourceUrlSenado?: string | null;
  temas?: Tema[];
  autores: Autor[];
  tramitacao: Tramitacao[];
  votacoes?: Votacao[];
}

export interface ProposicoesFiltros {
  anos: number[];
}

/**
 * @nullable
 */
export type ResultadosBuscaProposicaoMatchExato = {
  tipo?: string;
  numero?: number;
  ano?: number;
} | null;

export interface ResultadosBusca {
  parlamentares: Parlamentar[];
  proposicoes: ProposicaoResumo[];
  votacoes: Votacao[];
  /** @nullable */
  proposicaoMatchExato?: ResultadosBuscaProposicaoMatchExato;
}

export interface PartidoDetalhe {
  sigla: string;
  /** @nullable */
  nomeOficial?: string | null;
  totalParlamentares: number;
  parlamentares: Parlamentar[];
  /** @nullable */
  fidelidadeMedia?: number | null;
  parlamentaresElegiveis?: number;
}

export type MetricasParlamentarPresenca = {
  presente?: number;
  total?: number;
  /** @nullable */
  percentual?: number | null;
};

export interface MetricasParlamentar {
  parlamentarId: string;
  presenca: MetricasParlamentarPresenca;
  gastosTotalGeral: string;
  gastosTotalRegistros: number;
  gastosTopCategorias: CategoriaGasto[];
  proposicoesAutoriaPrimaria: number;
}

export interface ConcordanciaPar {
  a: string;
  b: string;
  votosEmComum: number;
  coincidentes: number;
  /** @nullable */
  percentual: number | null;
}

export interface CompararResult {
  parlamentares: Parlamentar[];
  metricas: MetricasParlamentar[];
  concordancia: ConcordanciaPar[];
  ano: number;
}

export type ListParlamentaresParams = {
  casa?: string;
  partido?: string;
  uf?: string;
};

export type ListVotacoesParams = {
  casa?: string;
  ano?: string;
  resultado?: string;
  somenteNominais?: string;
};

export type ListProposicoesParams = {
  tipo?: string;
  ano?: string;
  situacao?: string;
};

export type BuscarParams = {
  q: string;
};

export type CompararParams = {
  ids: string;
};