# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Brasil a Vera API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parlamentares:
    get:
      operationId: listParlamentares
      tags: [parlamentares]
      summary: List parliamentarians with optional filters
      parameters:
        - name: casa
          in: query
          schema:
            type: string
        - name: partido
          in: query
          schema:
            type: string
        - name: uf
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of parlamentares
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Parlamentar"
...
  /parlamentares/{id}:
    get:
      operationId: getParlamentar
      tags: [parlamentares]
      summary: Get a single parliamentarian by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Parlamentar detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParlamentarDetalhe"
        "404":
          description: Not found
  /parlamentares/{id}/votos-recentes:
    get:
      operationId: getVotosRecentes
      tags: [parlamentares]
      summary: Get recent votes for a parliamentarian
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recent votes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VotoRecente"
  /parlamentares/{id}/proposicoes-autoradas:
    get:
      operationId: getProposicoesAutoradas
      tags: [parlamentares]
      summary: Get bills authored by a parliamentarian
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Authored proposicoes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProposicaoResumo"
  /parlamentares/{id}/gastos-resumo:
    get:
      operationId: getGastosResumo
      tags: [parlamentares]
      summary: Get expense summary for a parliamentarian
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Expense summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GastosResumo"
  /parlamentares/{id}/alinhamento:
    get:
      operationId: getAlinhamento
      tags: [parlamentares]
      summary: Get party alignment stats for a parliamentarian
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alignment data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlinhamentoResult"
  /parlamentares/{id}/afinidade-voto:
    get:
      operationId: getAfinidadeVoto
      tags: [parlamentares]
      summary: Get top vote affinity partners for a parliamentarian
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Vote affinity partners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AfinidadeVoto"
...
  /partidos/{sigla}:
    get:
      operationId: getPartido
      tags: [partidos]
      summary: Get partido overview
      parameters:
        - name: sigla
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Partido detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartidoDetalhe"
        "404":
          description: Not found
  /comparar:
    get:
      operationId: comparar
      tags: [parlamentares]
      summary: Compare 2-3 parlamentares
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comparison result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompararResult"
        "400":
          description: Bad request
components:
  schemas:
...
    Parlamentar:
      type: object
      required: [id, nome, casa, partidoSigla, uf, legislatura, trustLevel, sourceUrl]
      properties:
        id:
          type: string
        nome:
          type: string
        casa:
          type: string
        partidoSigla:
          type: string
        partidoNome:
          type: string
        uf:
          type: string
        urlFoto:
          type: ["string", "null"]
        legislatura:
          type: integer
        trustLevel:
          type: string
        sourceUrl:
          type: string
    ParlamentarDetalhe:
      type: object
      required: [id, nome, casa, partidoSigla, uf, legislatura, trustLevel, sourceUrl]
      properties:
        id:
          type: string
        sourceId:
          type: string
        nome:
          type: string
        nomeCivil:
          type: ["string", "null"]
        casa:
          type: string
        partidoSigla:
          type: string
        partidoNome:
          type: string
        uf:
          type: string
        urlFoto:
          type: ["string", "null"]
        situacaoMandato:
          type: string
        legislatura:
          type: integer
        trustLevel:
          type: string
        sourceUrl:
          type: string
...
    VotoRecente:
      type: object
      required: [voto, votacaoId, dataHora, descricao, orgao, casa, aprovada]
      properties:
        voto:
          type: string
        votacaoId:
          type: string
        votacaoSourceId:
          type: ["string", "null"]
        dataHora:
          type: string
        descricao:
          type: string
        orgao:
          type: string
        casa:
          type: string
        aprovada:
          type: boolean
    ProposicaoResumo:
      type: object
      required: [id, tipo, numero, ano, ementa, situacao]
      properties:
        id:
          type: string
        tipo:
          type: string
        numero:
          type: integer
        ano:
          type: integer
        ementa:
          type: string
        situacao:
          type: string
    GastosResumo:
      type: object
      required: [totalGeral, totalRegistros, topCategorias, ano]
      properties:
        totalGeral:
          type: string
        totalRegistros:
          type: integer
        topCategorias:
          type: array
          items:
            $ref: "#/components/schemas/CategoriaGasto"
        ano:
          type: integer
    CategoriaGasto:
      type: object
      required: [categoriaDescricao, total, n]
      properties:
        categoriaDescricao:
          type: string
        total:
          type: string
        n:
          type: integer
    AlinhamentoResult:
      type: object
      required: [total, alinhados, divergentes, amostraInsuficiente, topDivergencias, topConvergencias]
      properties:
        partidoSigla:
          type: ["string", "null"]
        percentual:
          type: ["number", "null"]
        total:
          type: integer
        alinhados:
          type: integer
        divergentes:
          type: integer
        amostraInsuficiente:
          type: boolean
        topDivergencias:
          type: array
          items:
            $ref: "#/components/schemas/VotacaoAlinhamento"
        topConvergencias:
          type: array
          items:
            $ref: "#/components/schemas/VotacaoAlinhamento"
    VotacaoAlinhamento:
      type: object
      required: [votacaoId, dataHora, descricao, voto, orientacao]
      properties:
        votacaoId:
          type: string
        dataHora:
          type: string
        descricao:
          type: string
        voto:
          type: string
        orientacao:
          type: string
    AfinidadeVoto:
      type: object
      required: [parlamentarId, nome, partidoSigla, uf, casa, votosCoincidentes, totalVotosEmComum, percentualAfinidade]
      properties:
        parlamentarId:
          type: string
        nome:
          type: string
        partidoSigla:
          type: string
        uf:
          type: string
        casa:
          type: string
        urlFoto:
          type: ["string", "null"]
        votosCoincidentes:
          type: integer
        totalVotosEmComum:
          type: integer
        percentualAfinidade:
          type: integer
...
    PartidoDetalhe:
      type: object
      required: [sigla, totalParlamentares, parlamentares]
      properties:
        sigla:
          type: string
        nomeOficial:
          type: ["string", "null"]
        totalParlamentares:
          type: integer
        parlamentares:
          type: array
          items:
            $ref: "#/components/schemas/Parlamentar"
        fidelidadeMedia:
          type: ["number", "null"]
        parlamentaresElegiveis:
          type: integer
    CompararResult:
      type: object
      required: [parlamentares, metricas, concordancia, ano]
      properties:
        parlamentares:
          type: array
          items:
            $ref: "#/components/schemas/Parlamentar"
        metricas:
          type: array
          items:
            $ref: "#/components/schemas/MetricasParlamentar"
        concordancia:
          type: array
          items:
            $ref: "#/components/schemas/ConcordanciaPar"
        ano:
          type: integer
    MetricasParlamentar:
      type: object
      required: [parlamentarId, presenca, gastosTotalGeral, gastosTotalRegistros, gastosTopCategorias, proposicoesAutoriaPrimaria]
      properties:
        parlamentarId:
          type: string
        presenca:
          type: object
          properties:
            presente:
              type: integer
            total:
              type: integer
            percentual:
              type: ["number", "null"]
        gastosTotalGeral:
          type: string
        gastosTotalRegistros:
          type: integer
        gastosTopCategorias:
          type: array
          items:
            $ref: "#/components/schemas/CategoriaGasto"
        proposicoesAutoriaPrimaria:
          type: integer
    ConcordanciaPar:
      type: object
      required: [a, b, votosEmComum, coincidentes, percentual]
      properties:
        a:
          type: string
        b:
          type: string
        votosEmComum:
          type: integer
        coincidentes:
          type: integer
        percentual:
          type: ["number", "null"]