# lib/api-zod/src/generated/types/alinhamentoResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Brasil a Vera API
 * OpenAPI spec version: 0.1.0
 */
import type { VotacaoAlinhamento } from "./votacaoAlinhamento";

export interface AlinhamentoResult {
  /** @nullable */
  partidoSigla?: string | null;
  /** @nullable */
  percentual?: number | null;
  total: number;
  alinhados: number;
  divergentes: number;
  amostraInsuficiente: boolean;
  topDivergencias: VotacaoAlinhamento[];
  topConvergencias: VotacaoAlinhamento[];
}